Move build pipeline into support modules
