+++ tests/test_build_refactor.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-import build
+from tools.build_support import paths, pipeline
 from tools.build_support.context import BuildContext
 
 
@@ -20,7 +20,7 @@ def make_context(tmp_path, system="Linux"):
 def test_resolve_output_dir_uses_dist_for_release_build(tmp_path):
     context = make_context(tmp_path)
 
-    assert build.resolve_output_dir(context, develop=False) == (
+    assert paths.resolve_output_dir(context, develop=False) == (
         tmp_path / "dist-portable" / "main"
     )
 
@@ -29,7 +29,7 @@ def test_resolve_output_dir_uses_linux_develop_path(monkeypatch, tmp_path):
     context = make_context(tmp_path, system="Linux")
     monkeypatch.setattr(Path, "home", lambda: tmp_path / "home")
 
-    assert build.resolve_output_dir(context, develop=True) == (
+    assert paths.resolve_output_dir(context, develop=True) == (
         tmp_path / "home" / ".config" / "TheAnimeScripter" / "TAS-Portable"
     )
 
@@ -41,31 +41,31 @@ def test_build_portable_runs_steps_in_order(monkeypatch, tmp_path):
     calls = []
 
     monkeypatch.setattr(
-        build, "download_portable_python", lambda ctx: calls.append("python")
+        pipeline, "download_portable_python", lambda ctx: calls.append("python")
     )
     monkeypatch.setattr(
-        build, "install_requirements", lambda ctx: calls.append("requirements")
+        pipeline, "install_requirements", lambda ctx: calls.append("requirements")
     )
     monkeypatch.setattr(
-        build,
+        pipeline,
         "bundle_files",
         lambda ctx, output_dir: calls.append(("bundle", output_dir)),
     )
     monkeypatch.setattr(
-        build,
+        pipeline,
         "move_extras",
         lambda ctx, output_dir: calls.append(("extras", output_dir)),
     )
     monkeypatch.setattr(
-        build,
+        pipeline,
         "cleanup_temp_files",
         lambda ctx, output_dir: calls.append(("cleanup", output_dir)),
     )
     monkeypatch.setattr(
-        build, "remove_portable_python", lambda ctx: calls.append("remove")
+        pipeline, "remove_portable_python", lambda ctx: calls.append("remove")
     )
 
-    output_dir = build.build_portable(context)
+    output_dir = pipeline.build_portable(context)
 
     assert output_dir == tmp_path / "dist-portable" / "main"
     assert calls == [
